refactor(structs): split Python msgspec struct definitions into submodules
This change refactors the Python msgspec.Struct definitions for TEI structures by splitting them into three focused internal helper modules: _structs_common.py, _structs_body.py, and _structs_events.py. The public-facing tei_rapporteur.structs module now imports these submodules to compose its API.

Benefits:
- Improved readability and maintainability by modularizing large structs.py
- Clear separation of common definitions, body content projections, and streaming event projections
- Embedded into native extension at compile time with coherent submodule structure

Also updated the Rust module registration code to load the new submodules and handle import errors consistently.
